ignore System.Text.Encodings.Web.dll

+++ Scripts/post_build.ps1
@@ -40,6 +40,10 @@ function Delete-Unused ($path, $config) {
     $target = "$path\Output\$config"
     $included = Get-ChildItem $target -Filter "*.dll"
     foreach ($i in $included){
+        if ($i.Name in ["System.Text.Encodings.Web.dll"]) {
+            # ignore some specific dll that seems to make issue
+            continue
+        }
         $deleteList = Get-ChildItem $target\Plugins -Filter $i.Name -Recurse | Where { $_.VersionInfo.FileVersion -eq $i.VersionInfo.FileVersion -And $_.Name -eq $i.Name } 
         $deleteList | ForEach-Object{ Write-Host Deleting duplicated $_.Name with version $_.VersionInfo.FileVersion at location $_.Directory.FullName }
         $deleteList | Remove-Item
